+++ tests/unit/bindings/cpp/testpkg-util.hpp
@@ -34,6 +34,8 @@ void appendComponentOpaqueData(std::vector<uint8_t>& pkg);
 
 // append FirmwareDeviceIdRecord
 void appendFirmwareDeviceIdRecord1(std::vector<uint8_t>& pkg);
+void appendFirmwareDeviceIdRecord1InvalidApplicableComponentOOB(
+    std::vector<uint8_t>& pkg);
 void appendFirmwareDeviceIdRecord2(std::vector<uint8_t>& pkg);
 void appendFirmwareDeviceIdRecord3(std::vector<uint8_t>& pkg);