+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Change categories:
 
 ### Fixed
 
+- bindings: cpp: PackageParser: Check for ApplicableComponents OOB index
 - dsp: pdr: Bound children read in `pldm_entity_association_pdr_extract()`
 - dsp: platform: Add const modifier to the pdr argument of
   pldm_platform_redfish_action_pdr_action_iter_init